fix: correctness bugs found in review (bit-identical)
All three numeric fixtures pass unchanged; the unit suite grows 75 -> 100
checks. Behaviour-preserving except where a code path was provably broken.

3EMA_SRSI_ATR never opened a position. EMA3_params has eight members and
the initializer supplied seven, so MAX_OPEN_TRADES_TO_TEST[i7] landed in
SRSIU and max_open_trades defaulted to 0 -- making the
`active_positions < MAX_OPEN_TRADES` guard always false. -Wextra flags
this now that the Makefile no longer suppresses it. F_3EMA_SRSI_ATR
initializes all eight and was unaffected.

calculate_result_metrics leaked NaN and inf. Zero trades divided by zero
for the win rate; zero drawdown divided by a zero ddc. Both reached the
printed banner and the score file, and an inf score would have beaten
every genuinely-measured candidate in the sweep's `score >` ranking. Both
degenerate cases now return finite, neutral metrics. gain_over_ddc is
computed once in ResultMetrics instead of being re-divided at each call
site.

RandomNumberGenerator discarded its own seeding: the constructor body
built a local mt19937 that shadowed the member, so the intended thread-id
mixing never took effect and the seed it logged was not the seed in use.

TA-Lib adapters validated inputs too late. run_ohlc_indicator compared
high/low/close sizes only after handing all three to TA-Lib, which would
already have read past the end of a short array. build_talib_output
trusted TA-Lib's (outBeg, outNbElement) window without bounds-checking it
against the output buffer.

INITIALIZE_DATA's start-index scan never stopped at the first match, so a
repeated timestamp would append a second entry and shift every later
pair's index, desynchronising start_indexes[ic] from PAIRS[ic]. The
wrapping while(true) around it never iterated.

TALIB_SuperTrend_dir_only was a 75-line verbatim copy of TALIB_SuperTrend
and is now a thin wrapper over it; both copies also built three series
(trend/long/short) that neither returned. atr_multi widens from int to
float so the usual fractional multipliers can be swept.

+++ custom_talib_wrapper.hh
@@ -44,10 +44,12 @@ std::vector<float> TALIB_STOCHRSI_not_averaged(const std::vector<float> &vals, c
 
 //////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////
 
+// atr_multi is a float so fractional multipliers (1.5, 2.5, ...) can be swept; it used
+// to be an int, which silently excluded the most common SuperTrend settings.
 SuperTrend TALIB_SuperTrend(const std::vector<float> &high, const std::vector<float> &low, const std::vector<float> &close,
-                            const int atr_window, const int atr_multi);
+                            const int atr_window, const float atr_multi);
 std::vector<float> TALIB_SuperTrend_dir_only(const std::vector<float> &high, const std::vector<float> &low, const std::vector<float> &close,
-                                             const int atr_window, const int atr_multi);
+                                             const int atr_window, const float atr_multi);
 //////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////
 std::vector<float> TALIB_AO(const std::vector<float> &high, const std::vector<float> &low,
                             const int fast, const int slow);
